Create interface for HandBrake class
