+++ auto-handbrake-core/src/main/java/com/willmolloy/handbrake/core/Cli.java
@@ -2,7 +2,6 @@ package com.willmolloy.handbrake.core;
 
 import static com.google.common.base.Preconditions.checkNotNull;
 
-import com.google.common.annotations.VisibleForTesting;
 import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
 import java.io.BufferedReader;
 import java.io.InputStream;
@@ -26,13 +25,8 @@ class Cli {
 
   private final Supplier<ProcessBuilder> processBuilderSupplier;
 
-  @VisibleForTesting
   Cli(Supplier<ProcessBuilder> processBuilderSupplier) {
     this.processBuilderSupplier = checkNotNull(processBuilderSupplier);
-  }
-
-  Cli() {
-    this(ProcessBuilder::new);
   }
 
   /**
